grid: separate animation part into its own file
